lucid-analytics/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic[email]
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
cryptography==42.0.0
python-multipart==0.0.6
